Time: 6 ms (30.86%), Space: 8.6 MB (28.19%) - LeetHub

+++ 0326-power-of-three/0326-power-of-three.cpp
@@ -4,7 +4,7 @@ public:
         if(n<=0){
             return false;
         }
-        long long temp= 1;
+        long int temp= 1;
         while(temp<=n){
             if(temp==n){
                 return true;
